requirements.txt
anyio==3.7.1
asyncclick==8.1.3.4
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
idna==3.4
pydantic==1.10.11
python-dotenv==1.0.0
python-http-client==3.3.7
python-kasa==0.5.2
requests==2.31.0
sendgrid==6.10.0
sniffio==1.3.0
starkbank-ecdsa==2.2.0
typing_extensions==4.7.1
urllib3==2.0.4
